requirements.txt
Flask
flask-cors
flask-pymongo
pymongo
mysql-connector-python
bcrypt
python-dotenv
gunicorn
openpyxl
pandas
openai